introduce adapter and handlers to order matching engine
