+++ nautilus_core/execution/src/models/fee.rs
@@ -125,7 +125,6 @@ impl FeeModel for MakerTakerFeeModel {
 
 #[cfg(test)]
 mod tests {
-    use nautilus_execution::models::fee::{FeeModel, FixedFeeModel, MakerTakerFeeModel};
     use nautilus_model::{
         enums::{LiquiditySide, OrderSide, OrderType},
         instruments::{stubs::audusd_sim, InstrumentAny},
@@ -137,6 +136,8 @@ mod tests {
     };
     use rstest::rstest;
     use rust_decimal::prelude::ToPrimitive;
+
+    use super::{FeeModel, FixedFeeModel, MakerTakerFeeModel};
 
     #[rstest]
     fn test_fixed_model_single_fill() {
